+++ packages/components/repeater/src/repeater.tsx
@@ -2,15 +2,19 @@ import React, { Fragment } from "react";
 
 type Props = {
   count: number;
-  children: ((idx: number) => React.ReactNode) | React.ReactNode;
+  children:
+    | ((_: { count: number; index: number }) => React.ReactNode)
+    | React.ReactNode;
 };
 
 export const Repeater = (props: Props) => {
   const { count, children } = props;
   const items = Array.from({ length: count }).fill(null);
-  return items.map((_, idx) => (
-    <Fragment key={idx}>
-      {typeof children === "function" ? children(idx) : children}
+  return items.map((_, index) => (
+    <Fragment key={index}>
+      {typeof children === "function"
+        ? children({ count, index: index })
+        : children}
     </Fragment>
   ));
 };