+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ui/storybook
 
+## 0.0.180
+
+### Patch Changes
+
+- @jamsr-ui/react@0.0.178
+
 ## 0.0.179
 
 ### Patch Changes