+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 0.1.160
+
+### Patch Changes
+
+- fix drawer,repeater
+  - @jamsr-ui/react@0.0.178
+
 ## 0.1.159
 
 ### Patch Changes