+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/menu
 
+## 0.0.83
+
+### Patch Changes
+
+- fix drawer,repeater
+
 ## 0.0.82
 
 ### Patch Changes