+++ packages/components/rating/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/rating
 
+## 0.0.22
+
+### Patch Changes
+
+- fix drawer,repeater
+- Updated dependencies
+  - @jamsr-ui/repeater@0.0.32
+
 ## 0.0.21
 
 ### Patch Changes